% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.r
\name{ms}
\alias{ms}
\title{Estimating the direct and indirect of the Multistate model.}
\usage{
ms(data, P.time)
}
\arguments{
\item{data}{data.frame(X1,X2,D,Z)}

\item{P.time}{interpolation time can be vector or scalar}
}
\description{
Estimating the direct and indirect of the Multistate model.
}
\examples{
#This function needs additonal packages "mstate"
#install.packages("mstate")
library(mstate)
data=meta.gen(500,theta_0=0.5,theta_1=0.5,L1=0.5,L2=0.5,L3=1,b01=1,b02=0,b03=0,cc=2,dd="uniform")
P.time=seq(0,1,by=0.01)
ans=ms(data,P.time)
plot(P.time,ans$DE,type="l",ylim=c(-0.5,0.5))
points(P.time,ans$IE,type="l",ylim=c(-0.5,0.5),col=2)
legend(0,0.45,c("direct effect","indirect effect"),col=1:2,lty=1)
}
\keyword{causal}
\keyword{frailty}
\keyword{inference,}
\keyword{model}
\keyword{risks,}
\keyword{semicompeting}
